requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
python-dotenv
requests
opencv-python
scikit-learn
matplotlib
seaborn
torch
torchvision
onnx
onnxruntime
pydantic
typing-extensions
httpx
python-multipart
